pyproject.toml
[tool.poetry]
name = "cloudantic"
version = "0.0.8"
description = ""
authors = ["Oscar Bahamonde <o.bahamonde@globant.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
boto3 = "^1.28.65"
pydantic = "1.10.12"
python-dotenv = "^1.0.0"
rich = "^13.6.0"
tenacity = "^8.2.3"
aiohttp = "^3.8.6"
aiohttp-sse = "^2.1.0"
openai = "^0.28.1"
tiktoken = "^0.5.1"
numpy = "^1.26.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.10.0"
isort = "^5.12.0"
pylint = "^3.0.1"
boto3-stubs = {version = "1.28.67", extras = ["cognito-idp", "dynamodb", "dynamodbstreams", "kinesis", "lambda", "s3"]}
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
